refactor(release): split T1.x callables from Vitest wrappers so orchestrator can import
Each T1.1/T1.2/T1.3 file previously combined the callable + inline describe/it
blocks, causing vitest imports to fire when run-tier-1.ts imported them outside
the vitest runner. Split into callable *.ts (no vitest deps) + sibling *.test.ts
that imports the callable. Narrow vitest.config.ts release include pattern from
*.ts to *.test.ts so bare callable files are not scanned as test files.
